# docs/swagger.json
{
  "openapi": "3.1.0",
  "info": {
    "version": "1.0.0",
    "title": "AquaTrack WebApp",
    "license": {
      "name": "Apache 2.0",
      "url": "<http://www.apache.org/licenses/LICENSE-2.0.html>"
    },
    "description": "This is a documentation of AquaTrack app\n"
  },
  "servers": [
    {
      "url": "https://aquatrack-webapp-backend.onrender.com"
    }
  ],
...
  "paths": {
...
    "/users/current": {
      "get": {
        "tags": [
          "user"
        ],
        "summary": "Get current user",
        "description": "Retrieves the current authenticated user's information",
        "operationId": "getCurrentAccauntController",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response with the current user's information.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "user": {
                      "type": "object",
                      "properties": {
                        "userID": {
                          "type": "string",
                          "description": "The user's ID.",
                          "example": "60d0fe4f5311236168a109ca"
                        },
                        "name": {
                          "type": "string",
                          "description": "The user's name.",
                          "example": "John Doe"
                        },
                        "email": {
                          "type": "string",
                          "description": "The user's email.",
                          "example": "john.doe@example.com"
                        },
                        "gender": {
                          "type": "string",
                          "description": "The user's gender.",
                          "example": "male"
                        },
                        "weight": {
                          "type": "number",
                          "description": "The user's weight in kg.",
                          "example": 75
                        },
                        "activeSportsTime": {
                          "type": "number",
                          "description": "The user's active sports time in minutes.",
                          "example": 30
                        },
                        "dailyWaterIntake": {
                          "type": "number",
                          "description": "The user's daily water intake in liters.",
                          "example": 2.5
                        },
                        "avatar": {
                          "type": "string",
                          "description": "URL of the user's avatar.",
                          "example": "https://example.com/avatar.jpg"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/401"
          }
        }
      }
    },
    "/users/:{userId}": {
      "get": {
        "tags": [
          "user"
        ],
        "summary": "Get user information",
        "operationId": "getUser",
        "parameters": [
          {
            "in": "path",
            "name": "userId",
            "required": true,
            "description": "User identifier",
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful retrieval of user information",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "Successfully found user!"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "userId": {
                          "type": "string"
                        },
                        "name": {
                          "type": "string"
                        },
                        "email": {
                          "type": "string"
                        },
                        "gender": {
                          "type": "string"
                        },
                        "weight": {
                          "type": "number"
                        },
                        "activeSportsTime": {
                          "type": "number"
                        },
                        "dailyWaterIntake": {
                          "type": "number"
                        },
                        "avatar": {
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "User not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 404
                    },
                    "message": {
                      "type": "string",
                      "example": "User not found"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      }
    },
    "/users/update": {
      "patch": {
        "tags": [
          "user"
        ],
        "summary": "Update user",
        "description": "Updates the current authenticated user's information",
        "operationId": "updateUserService",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "description": "The user's name.",
                    "example": "John Doe"
                  },
                  "email": {
                    "type": "string",
                    "description": "The user's email.",
                    "example": "john.doe@example.com"
                  },
                  "gender": {
                    "type": "string",
                    "description": "The user's gender.",
                    "example": "male"
                  },
                  "weight": {
                    "type": "number",
                    "description": "The user's weight in kg.",
                    "example": 75
                  },
                  "activeSportsTime": {
                    "type": "number",
                    "description": "The user's active sports time in minutes.",
                    "example": 30
                  },
                  "dailyWaterIntake": {
                    "type": "number",
                    "description": "The user's daily water intake in liters.",
                    "example": 2.5
                  },
                  "avatar": {
                    "type": "string",
                    "format": "binary",
                    "description": "The user's avatar image file."
                  }
                }
              },
              "encoding": {
                "photo": {
                  "contentType": "image/*"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response with the updated userinformation.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "User updated successfully"
                    },
                    "data": {
                      "type": "object",
                      "$ref": "#/components/schemas/user"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      }
    },
    "/users/refresh-tokens": {
      "post": {
        "tags": [
          "user"
        ],
        "summary": "Refresh access token",
        "description": "Refresh access token",
        "operationId": "refreshSessionService",
        "security": [],
        "responses": {
          "200": {
            "description": "Successfully refreshed a session!",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "Successfully refreshed a session!"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "accessToken": {
                          "type": "string",
                          "example": "264564cnkvjguTLJFRS934"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "401": {
            "$ref": "#/components/responses/401"
          }
        }
      }
    },
    "/users/logout": {
      "post": {
        "tags": [
          "user"
        ],
        "summary": "Logout user",
        "operationId": "logoutUserService",
        "description": "logout user",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "204": {
            "description": "No Content - User was successfully logged out"
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    },
    "schemas": {
...
      "user": {
        "type": "object",
        "required": [
          "userId",
          "email",
          "password"
        ],
        "properties": {
          "userID": {
            "type": "string",
            "description": "The user's ID.",
            "example": "60d0fe4f5311236168a109ca"
          },
          "name": {
            "type": "string",
            "description": "The user's name.",
            "example": "John Doe"
          },
          "email": {
            "type": "string",
            "description": "The user's email.",
            "example": "john.doe@example.com"
          },
          "gender": {
            "type": "string",
            "description": "The user's gender.",
            "example": "male"
          },
          "weight": {
            "type": "number",
            "description": "The user's weight in kg.",
            "example": 75
          },
          "activeSportsTime": {
            "type": "number",
            "description": "The user's active sports time in minutes.",
            "example": 30
          },
          "dailyWaterIntake": {
            "type": "number",
            "description": "The user's daily water intake in liters.",
            "example": 2.5
          },
          "avatar": {
            "type": "string",
            "description": "URL of the user's avatar.",
            "example": "https://example.com/avatar.jpg"
          }
        }
      }
    },
    "responses": {
      "400": {
        "description": "The input data does not match the validation scheme!",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "required": [
                "status",
                "message"
              ],
              "properties": {
                "status": {
                  "type": "integer",
                  "example": 400
                },
                "message": {
                  "type": "string",
                  "example": "The input data does not match the validation scheme!"
                }
              }
            }
          }
        }
      },
      "401": {
        "description": "Error, user is not authorized",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "required": [
                "status",
                "message",
                "data"
              ],
              "properties": {
                "status": {
                  "type": "integer",
                  "example": 401
                },
                "message": {
                  "type": "string",
                  "example": "Error, user is not authorized!"
                },
                "data": {
                  "type": "object",
                  "required": [
                    "message"
                  ],
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Access token expired!"
                    }
                  }
                }
              }
            }
          }
        }
      },
...
      "500": {
        "description": "Server Error",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Internal Server Error"
                }
              }
            }
          }
        }
      }
    }
  }
}